Create tokenizer.txt
Add this text file in the same directory as the tokenizer program for it to run properly. Any name will suffice, just be sure you use that name in the command line arguments to open the file. Feel free to add any more characters to see what outputs you get